Refactor after last laborator, change -> one file for each class-> change also the logic, did not see any logic to add the names of the animals inside the constructor
