+++ animal.py
@@ -1,0 +1,5 @@
+class Animal:
+    sound = ""
+
+    def set_category(self, sound):
+        self.sound = sound
