+++ animal_types.py
@@ -1,0 +1,11 @@
+from animal import Animal
+
+class Pig(Animal):
+    sound = 'quizquiz'
+
+
+class Cat(Animal):
+    sound='meow meow'
+
+class Dog(Animal):
+    sound='bark bark'
